+++ NEWS.md
@@ -1,6 +1,9 @@
 
 # rstudioapi 0.14 (UNRELEASED)
 
+* `getSourceEditorContext()` gains the `id` argument, to be used to request
+  the editor context for a document with an already-known ID. (#251)
+  
 * Added `documentOpen()`, for opening a document in RStudio and optionally
   navigating the cursor to a particular point in the file. The method is
   synchronous and returns the document ID upon completion.
@@ -12,9 +15,11 @@
   `unregisterCommandCallback`, used to execute a callback after an IDE command
   is executed.
   
+
 # rstudioapi 0.13
 
 * Fixed an issue where `rstudioapi::insertText()` would fail. (#208)
+
 
 # rstudioapi 0.12
 
@@ -26,6 +31,7 @@
 
 * Added `registerChunkExecCallback` and `unregisterChunkExecCallback`, used to
   execute a callback after a chunk is ran.
+
 
 # rstudioapi 0.11
 
@@ -50,15 +56,18 @@
 * Added `userIdentity` and `systemUsername`, used to retrieve information about
   the current user.
 
+
 # rstudioapi 0.10
 
 * Added the parameters `echo` and `focus` to `sendToConsole()`.
+
 
 # rstudioapi 0.9
 
 * Added functions for displaying jobs in RStudio's Jobs pane: `jobAdd()`, `jobRemove()`, etc.
 
 * Added `translateLocalUrl()`, for translating localhost URLs to externally addressable ones on RStudio Server.
+
 
 # rstudioapi 0.8
 
@@ -91,11 +100,13 @@
   `terminalRunning()`, `terminalKill()`, `terminalSend()`, `terminalExecute()`,
   and `terminalExitCode()`.
 
+
 # rstudioapi 0.6
 
 * Add sendToConsole function
 
 * Add APIs for setting cursor position in document
+
 
 # rstudioapi 0.5
 
@@ -103,13 +114,16 @@
 
 * Add getActiveProject function
 
+
 # rstudioapi 0.4
 
 * Add API methods for interacting with a document open in RStudio: 'insertText()', 'modifyRange()' and 'getActiveDocumentContext()'.
 
+
 # rstudioapi 0.3
 
 * Add stub and documentation for sourceMarker function
+
 
 # rstudioapi 0.2
 
@@ -117,7 +131,7 @@
 
 * Stubs and documentation for versionInfo, previewRd, and viewer functions
 
+
 # rstudioapi 0.1
 
-Initial release to CRAN
-
+* Initial release to CRAN